chore(release): promote private draft into develop
Why: publish the current validated draft lane to public develop while preserving the repo's squash-based promotion history.

What: squash-promote private/draft at 0bc2be7, including RPC decoding hardening, subscription lifecycle cleanup, internal naming simplifications, catalog validation, and reconnect/test updates.

Testing: swift test --no-parallel; swift test --filter ClientCancellationValidator/subscribeTaskCancellationDoesNotEmitLateRequest; local artifact guard passed.

+++ Sources/SwiftFulcrum/RPC/Method/Core/Method.Blockchain+UTXO.swift
@@ -4,6 +4,6 @@ import Foundation
 
 extension SwiftFulcrum.RPC.Method.Blockchain {
     enum UTXO: Sendable {
-        case getInfo(transactionHash: String, outputIndex: UInt16)
+        case getInfo(transactionHash: String, outputIndex: UInt)
     }
 }
